refactor : API 전달되는 객체 dto 패키지로 이동 & mask stock 에 정의된 함수를 확장함수로 뺌
