+++ app/src/main/java/com/survivalcoding/maskinfo/data/dto/ResultGetMaskStock.kt
@@ -1,0 +1,6 @@
+package com.survivalcoding.maskinfo.data.dto
+
+data class ResultGetMaskStock(
+    val count: Int,
+    val stores: List<Store>,
+)
