--- app/src/main/java/com/survivalcoding/maskinfo/data/model/ResultGetMaskStock.kt
@@ -1,6 +1,0 @@
-package com.survivalcoding.maskinfo.data.model
-
-data class ResultGetMaskStock (
-    val count: Int = 222,
-    val stores: List<Store>,
-)
