+++ app/src/main/res/layout/mask_stock_item.xml
@@ -6,6 +6,8 @@
 
         <import type="androidx.core.content.ContextCompat" />
 
+        <import type="com.survivalcoding.maskinfo.data.extension.MaskStockExtKt" />
+
         <variable
             name="maskStock"
             type="com.survivalcoding.maskinfo.data.model.MaskStock" />
@@ -47,9 +49,9 @@
             android:layout_width="100dp"
             android:layout_height="wrap_content"
             android:layout_marginTop="5dp"
-            android:text="@{maskStock.amountString()}"
+            android:text="@{MaskStockExtKt.amountString(maskStock)}"
             android:textAlignment="center"
-            android:textColor="@{ContextCompat.getColor(context, maskStock.statusColor())}"
+            android:textColor="@{ContextCompat.getColor(context, MaskStockExtKt.statusColor(maskStock))}"
             android:textSize="16sp"
             app:layout_constraintEnd_toEndOf="@+id/stateTextView"
             app:layout_constraintTop_toBottomOf="@+id/stateTextView" />
@@ -60,9 +62,9 @@
             android:layout_height="wrap_content"
             android:layout_marginTop="15dp"
             android:layout_marginEnd="7dp"
-            android:text="@{maskStock.statusString()}"
+            android:text="@{MaskStockExtKt.statusString(maskStock)}"
             android:textAlignment="center"
-            android:textColor="@{ContextCompat.getColor(context, maskStock.statusColor())}"
+            android:textColor="@{ContextCompat.getColor(context, MaskStockExtKt.statusColor(maskStock))}"
             android:textSize="18sp"
             android:textStyle="bold"
             app:layout_constraintEnd_toEndOf="parent"
